Module Capnp_rpc_lwt

Lwt compatibility wrappers.

This module wraps Capnp_rpc to provide the old Lwt API, to make upgrading easier. See https://github.com/mirage/capnp-rpc/blob/master/CHANGES.md for migration instructions.

module Segment = Capnp_rpc.Segment

module Message = Capnp_rpc.Message

module Slice = Capnp_rpc.Slice

module StructStorage = Capnp_rpc.StructStorage

module ListStorage = Capnp_rpc.ListStorage

module Object = Capnp_rpc.Object

module StructRef : sig ... end

A promise for a response structure. You can use the generated _get_pipelined functions on a promise to get a promise for a capability inside the promise, and then pipeline messages to that promise without waiting for the response struct to arrive.

module Capability : sig ... end

A capability is a reference to an object, or to a promise for an object. You can invoke methods on a capability even while it is still only a promise.

module Sturdy_ref : sig ... end

module Service : sig ... end

Functions for service implementors.

module Persistence : sig ... end