Module Capability.Request

type 'a t

An 'a t is a builder for the out-going request's payload.

val create : ?message_size:int -> (Capnp.Message.rw Slice.t -> 'a) -> 'a t * 'a

create init is a fresh request payload and contents builder. Use one of the generated init_pointer functions for init.

parameter message_size
An estimate of the size of the payload. If this is too small, additional segments will be allocated automatically, but this is less efficient than getting the size right to start with.

val create_no_args : unit -> 'a t

create_no_args () is a payload with no content.

val release : 'a t -> unit

Clear the exported refs, dropping their ref-counts. This is called automatically when you send a message, but you might need it if you decide to abort.