Module Service.Response

type 'b t

An 'a t is a builder for the out-going response.

val create : ?message_size:int -> (Capnp.Message.rw Slice.t -> 'a) -> 'a t * 'a

create init is a fresh response and contents builder. Use one of the generated init_pointer functions for init.

val create_empty : unit -> 'a t

empty () is an empty response.

val release : 'a t -> unit

Clear the exported refs, dropping their ref-counts. This is called automatically when you send a message, but you might need it if you decide to abort.