Module Msg.Request

include Capnp_rpc.S.WIRE_PAYLOAD with type path := Path.t and type t = request msg

type t = request msg

A message payload. This is typically a byte array of some kind, plus a way of attaching capabilities.

type path

val pp : t Fmt.t

val cap_index : t -> path -> int option

cap_index msg path is the capability index at path in the message (i.e the index of the capability in the attachments table).

val attachments : t -> Capnp_rpc.S.attachments

val with_attachments : Capnp_rpc.S.attachments -> t -> t

val writable : t -> Capnp_rpc_lwt__.Schema.Builder.Call.t

We're about to transmit this message and we need to fill in the target and CapDescriptor table. If the message is read-only (e.g. we received it over the network and are forwarding it) then make a copy. The copy has no CapDescriptor table and no attachments.

val readable : t -> Capnp_rpc_lwt__.Schema.Reader.Call.t

val of_builder : Capnp_rpc_lwt__.Schema.Builder.Call.t -> t

val of_reader : Capnp_rpc_lwt__.Schema.Reader.Call.t -> t