Module type Capnp_rpc_lwt__S.VAT_NETWORK

type +'a capability

An 'a capability is a capability reference to a service of type 'a.

type flow

A bi-directional byte-stream.

type restorer

A function for restoring persistent capabilities from sturdy ref service IDs.

type +'a sturdy_ref

An off-line (persistent) capability.

type service_id

A (secret) token that identifies a persistent service within a vat and grants access to it.

module Network : NETWORK

module CapTP : sig ... end

module Vat : sig ... end