Module Capnp_rpc_lwt

Cap'n Proto RPC using the Cap'n Proto serialisation and Lwt for concurrency.

include module type of Capnp.BytesMessage

module Segment : sig ... end

module Message : sig ... end

module Slice : sig ... end

module StructStorage : sig ... end

module ListStorage : sig ... end

module Object : sig ... end

module StructRef : sig ... end

A promise for a response structure. You can use the generated _get_pipelined functions on a promise to get a promise for a capability inside the promise, and then pipeline messages to that promise without waiting for the response struct to arrive.

module Capability : sig ... end

A capability is a reference to an object, or to a promise for an object. You can invoke methods on a capability even while it is still only a promise.

module Sturdy_ref : sig ... end

module Service : sig ... end

Functions for service implementors.

module Persistence : sig ... end