Module Capnp_rpc_lwt.StructRef

A promise for a response structure. You can use the generated _get_pipelined functions on a promise to get a promise for a capability inside the promise, and then pipeline messages to that promise without waiting for the response struct to arrive.

type 'a t

An 'a t is a reference to a response message (that may not have arrived yet) with content type 'a.

val inc_ref : 'a t -> unit

inc_ref t increases the reference count on t by one.

val dec_ref : 'a t -> unit

dec_ref t reduces the reference count on t by one. When the count reaches zero, this result must never be used again. If the results have not yet arrived when the count reaches zero, we send a cancellation request (which may or may not succeed). As soon as the results are available, they are released.